Introduce a new function that will check the return of factory().
The functions in this code makes many assumptions about what methods
and capabilities are available on factory objects.
In order to guarantee that everything works properly, a constraint is
being documented on factories that they produce a mutable mapping.
With this constraint, we can make assumptions in dicttoolz about what we
can do with these mappings, namely that we can call .update() and assign
key/values.

This allows any subclass of MutableMapping ABC to work with dicttoolz.